初始提交

+++ database/migrations/2018_10_06_142358_create_advertisements_table.php
@@ -18,9 +18,9 @@ class CreateAdvertisementsTable extends Migration
             $table->string('title')->comment('广告标题');
             $table->string('img')->nullable()->comment('广告文件');
             $table->string('img_uri')->nullable()->comment('广告链接');
-            $table->decimal('ad_expenses', 10,2)->default(0)->comment('广告经费');
+            $table->integer('ad_expenses')->default(0)->comment('广告经费');
             $table->integer('divided_count')->default(0)->comment('被分润次数');
-            $table->decimal('divided_amount', 10,2)->default(0)->comment('被分润金额');
+            $table->integer('divided_amount')->default(0)->comment('被分润金额');
             $table->boolean('status')->comment('状态');
             $table->timestamps();
         });
